test: Ajusta testes para as modificações feitas na autenticação de usuário
